move logdouble::sum to a .cpp so we can use it in the covariate model
